Only show questions that belong to the quiz you were looking at
Before, 'show questions' button would show every question on the
database. I wanted to make it so that you would only see the questions
that were made for that quiz.
I exported 'quizid' from quiz-show.js and imported it in quiz-show
This quizid is always the quiz id of the most recent quiz the user is
viewing. Then in quiz-index, I run a check and only show questions that
have a quiz id that matches the quizid.